+++ src/lib/errors.ts
@@ -1,0 +1,20 @@
+export class UserFacingError extends Error {
+  constructor(message: string) {
+    super(message)
+    this.name = 'UserFacingError'
+  }
+}
+
+export class VersionIncompatibleError extends UserFacingError {
+  constructor(message: string) {
+    super(message)
+    this.name = 'VersionIncompatibleError'
+  }
+}
+
+export class ServerTooOldError extends UserFacingError {
+  constructor(message: string) {
+    super(message)
+    this.name = 'ServerTooOldError'
+  }
+}